your_gingivity/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
tensorflow-cpu==2.15.0 # Lightweight CPU version
opencv-python-headless==4.8.1.78
jinja2==3.1.2
python-dotenv==1.0.0
aiofiles==23.2.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2


 
